backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
google-generativeai
langchain-google-genai
python-dotenv
langchain-text-splitters
pydantic